requirements.txt
# Core dependencies
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Claude Agent SDK
claude-agent-sdk>=0.1.0
